Time: 9 ms (57.47%) | Memory: 18.7 MB (39.77%) - LeetSync

+++ 229-majority-element-ii/majority-element-ii.cpp
@@ -36,8 +36,8 @@ public:
             }
         }
 
-        if(cnt1 >= int(n/3)+1) ans.push_back(ele1);
-        if(cnt2 >= int(n/3)+1) ans.push_back(ele2);
+        if(cnt1 >= (n/3)+1) ans.push_back(ele1);
+        if(cnt2 >= (n/3)+1) ans.push_back(ele2);
 
         return ans;
     }
